[refactor] VendingMachine의 기능을 등록과 계산으로 분리
- VendingMachineRegister : 자판기에 필요한 것을 등록하는 역할

- VendingMachineCalculator : 자판기에 필요한 계산을 하는 역할
