+++ src/main/java/vendingmachine/Application.java
@@ -1,10 +1,10 @@
 package vendingmachine;
 
-import vendingmachine.controller.VendingMachine;
+import vendingmachine.controller.VendingMachineController;
 
 public class Application {
     public static void main(String[] args) {
-        VendingMachine vendingMachine = new VendingMachine();
-        vendingMachine.run();
+        VendingMachineController vendingMachineController = new VendingMachineController();
+        vendingMachineController.run();
     }
 }
